Mommy, the Data Man is here!
DataMan being an abbreviation of "Data Manager". Eventually it'll be used to retrieve products from our text files. For now, it's just going to generate some so that we can test other things. It'll also be used to create receipts. The Data Man is so helpful!

(Oh, and also we finished out the Book and Game classes, which we forgot to commit last time.)